Refactor workflow and update requirements
