+++ .github/workflows/tests.yaml
@@ -3,10 +3,12 @@ name: Unit Tests
 on:
   push:
     branches:
-      -  [ master, main ]
+      -  master
+      -  main
   pull_request:
     branches:
-      -  [ master, main ]
+      -  master
+      -  main
 
 jobs:
   test:
@@ -40,6 +42,7 @@ jobs:
         run: |
           python -m pip install -U pip setuptools wheel
           pip install -r requirements.txt
+          pip install -r requirements_tests.txt
           
       - name: Run tests & coverage
         run: |
