+++ requirements.txt
@@ -13,5 +13,3 @@ fastapi==0.115.6
 uvicorn==0.34.0
 ruff==0.8.6
 hydra-core==1.3.2
-pytest==8.3.4
-coverage==7.6.10
